binance_scalper/requirements.txt
#requirements.txt

python-binance
python-dotenv
